created new styles/ directory and moved all sale css files to new directory
