+++ src/components/sale/card.js
@@ -1,5 +1,6 @@
 import React from "react";
-import "./card.css"
+// import "./card.css"
+import "../../styles/sales/card.css"
 
 const SaleCard = (props) => {
