--- src/components/sale/editForm.css
@@ -1,3 +1,0 @@
-.modal--input {
-    min-height: 35px;
-}
