+++ src/components/sale/saleDetailsModal.js
@@ -1,5 +1,6 @@
 import React from "react";
-import "./card.css"
+// import "./card.css"
+import "../../styles/sales/card.css"
 import "../employee/list.css"
 import "../employee/card.css"
 import NumberFormat from 'react-number-format';
