+++ src/styles/sales/detail.css
@@ -1,4 +1,4 @@
-/* .sale-details--container {
+.sale-details--container {
     border: 3px solid black;
     border-radius: 8px;
     max-width: 270px;
@@ -17,4 +17,4 @@ p {
     font-family: 'Roboto', sans-serif;
     font-weight: 300;
     font-size: 16px;    
-} */
+}
