+++ src/styles/sales/sale.css
@@ -1,9 +1,6 @@
-/* .salesContainer {
+.salesContainer {
     padding: 10px;
     min-width: 600px;
     max-width: 600px;
     max-height: 250px;
 }
-
-
- */
